refactor(library): Extract component to represent Scottish Parliament elections
